refactor: Client 모듈 리팩토링
