+++ lovebird-api/src/main/kotlin/com/lovebird/api/provider/oauth/KakaoAuthProvider.kt
@@ -3,7 +3,7 @@ package com.lovebird.api.provider.oauth
 import com.lovebird.api.dto.param.user.OAuthParam
 import com.lovebird.api.provider.PublicKeyProvider
 import com.lovebird.api.validator.JwtValidator
-import com.lovebird.webClient.client.KakaoAuthClient
+import com.lovebird.client.web.KakaoAuthClient
 import io.jsonwebtoken.Claims
 import org.springframework.stereotype.Component
