+++ lovebird-client/src/main/kotlin/com/lovebird/client/web/KakaoAuthClient.kt
@@ -1,7 +1,7 @@
-package com.lovebird.webClient.client
+package com.lovebird.client.web
 
-import com.lovebird.webClient.client.properties.KakaoProperties
-import com.lovebird.webClient.vo.key.OidcPublicKeyList
+import com.lovebird.client.vo.key.OidcPublicKeyList
+import com.lovebird.client.web.properties.KakaoProperties
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.client.WebClient
