+++ lovebird-client/src/main/kotlin/com/lovebird/client/web/PublicKeyClient.kt
@@ -1,0 +1,8 @@
+package com.lovebird.client.web
+
+import com.lovebird.client.vo.key.OidcPublicKeyList
+
+interface PublicKeyClient {
+
+	fun getPublicKeys(): OidcPublicKeyList
+}
