+++ lovebird-client/src/main/kotlin/com/lovebird/client/web/dto/response/NaverUserInfoClientResponse.kt
@@ -1,4 +1,4 @@
-package com.lovebird.webClient.client.dto.response
+package com.lovebird.client.web.dto.response
 
 data class NaverUserInfoClientResponse(
 	val resultCode: String,
