+++ lovebird-client/src/main/kotlin/com/lovebird/client/web/properties/GoogleProperties.kt
@@ -1,4 +1,4 @@
-package com.lovebird.webClient.client.properties
+package com.lovebird.client.web.properties
 
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdTokenVerifier
 import com.google.api.client.http.javanet.NetHttpTransport
