+++ lovebird-client/src/main/resources/client-prod.yml
