--- lovebird-external/web-client/src/main/kotlin/com/lovebird/webClient/client/PublicKeyClient.kt
@@ -1,8 +1,0 @@
-package com.lovebird.webClient.client
-
-import com.lovebird.webClient.vo.key.OidcPublicKeyList
-
-interface PublicKeyClient {
-
-	fun getPublicKeys(): OidcPublicKeyList
-}
